move yaml_dataset_loader logic to dataset_definition
